Functions to read individual maps.

These are worker functions to read a specific file from the remote sensing data collection.

read_i_ice(i, files)

read_i_sst(i, files)

read_i_u(i, files)

read_i_v(i, files)

read_i_mag(i, files)

read_i_chl(i, files)

Arguments

i

the index of the file

files

the table of file names and dates

Value

[raster::RasterLayer()]

Details

These functions encapsulate specific optimizations for extracting time series maps for the study. The `i-th` index is one of the rows of the `files`` data frame.

The specific data sets in use are:

`read_i_ice` reads the NSIDC SMMR-SSM/I Nasateam daily sea ice concentration, Passive-microwave estimates of sea ice concentration at 25km spatial resolution. Daily and monthly resolution, available from 1-Oct-1978 to present. http://nsidc.org/data/nsidc-0051.html

`read_i_sst` reads the NOAA Optimum Interpolation 1/4 Degree Daily Sea Surface Temperature Analysis, Sea surface temperature at 0.25 degree daily resolution, from 1-Sep-1981 to present. Richard W. Reynolds, Viva F. Banzon, and NOAA CDR Program (2008): NOAA Optimum Interpolation 1/4 Degree Daily Sea Surface Temperature (OISST) Analysis, Version 2. [southern hemisphere, 1981-2016]. NOAA National Climatic Data Center. doi:10.7289/V5SQ8XB5 [November, 2016] http://www.ngdc.noaa.gov/docucomp/page?xml=NOAA/NESDIS/NCDC/Geoportal/iso/xml/C00844.xml&view=getDataView&header=none

`read_i_u`, `read_i_v` and `read_i_mag` read 0.25 degree Ssalto/Duacs gridded absolute geostrophic velocities, Absolute geostrophic velocities computed from gridded sea surface heights above geoid from satellite altimetry (delayed and near-real-time products). http://www.aviso.altimetry.fr/en/data/products/sea-surface-height-products/global/madt.html. The altimeter products were produced by Ssalto/Duacs and distributed by Aviso, with support from Cnes (http://www.aviso.altimetry.fr/duacs/

`read_i_chl` read from each of Oceandata VIIRS, MODISA and SeaWiFS Level-3 mapped 32-day 9km chlorophyll-a, rolling 32-day composite remote-sensing chlorophyll-a from the VIIRS, MODISA and SeaWiFS satellites at 9km spatial resolution". SeaWiFS is 1997-2002, MODISA is 2002-2012, and VIIRS is 2012-2016. http://oceancolor.gsfc.nasa.gov

Underlying each function is the infrastructure provided by an installation of `raadtools` and `raadsync`: https://github.com/AustralianAntarcticDivision/raadtools https://github.com/AustralianAntarcticDataCentre/raadsync